Import register_config
* Separated config store implementation from the actual configurations
* Refactored unncessary subclassing of `TargetConf`
* Currently ignoring hydra-run 'outputs' directory
